NoDrop Attunable items show up Attunable.
Items which are both nodrop and attunable appear as light blue, which is attunable. 

I feel that since they are already attuned, they should show up as no drop instead. 
In order to make this change we need to hold on to ItemPtr a bit longer and there-fore it has to be sent to the function SetItemBG instead of just the ItemDefinition*
The reason for holding onto it is so we have access to pItem->NoDropFlag which is applied in the event of an item being set as NoDrop that could be either droppable or No Drop.